requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.30.0
sqlalchemy~=2.0.0
pymssql~=2.3.0
pydantic~=2.10.0
pydantic-settings~=2.6.0
httpx~=0.27.0
python-dotenv~=1.0.0
trustme~=1.1.0
pillow~=10.4.0
itsdangerous~=2.2.0